Fix a big problem that diagonal search results are overwritten by orthogonal searches
If diagonal search was done, orthogonal searches are skipped in that pixel,
however the vertical search still can be done in the right neighboring pixel
and overwrite the orthogonal search results. Such behavior badly spoiles the
effectiveness of the diagonal processing. Maybe there is the same problem in
the original implementation as well (not confirmed, though).